requirements.txt
autopep8==1.5.7
hubspot3==3.2.50
pycodestyle==2.7.0
python-dotenv
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
